2019.06.03 All tests except ITK transformation matrix passed.

+++ imagedata/axis.py
@@ -75,7 +75,7 @@ class UniformLengthAxis(UniformAxis):
             start = self.start + (item.start or 0) * self.step
             stop = self.start + (item.stop or self.n) * self.step
             step = (item.step or 1) * self.step
-            n = round((stop - start) / step)
+            n = int(round((stop - start) / step))
             n = min(self.n, n)
         #logging.debug('UniformLengthAxis: slice %d,%d,%d' % (start,stop,step))
         return UniformLengthAxis(self.name, start, n, step)
